+++ packages/std/src/imports.rs
@@ -156,8 +156,15 @@ impl ExternalApi {
 }
 
 impl Api for ExternalApi {
-    fn addr_validate(&self, human: &str) -> StdResult<Addr> {
-        let source = build_region(human.as_bytes());
+    fn addr_validate(&self, input: &str) -> StdResult<Addr> {
+        let input_bytes = input.as_bytes();
+        if input_bytes.len() > 90 {
+            // See MAX_LENGTH_HUMAN_ADDRESS in the VM.
+            // In this case, the VM will refuse to read the input from the contract.
+            // Stop here to allow handling the error in the contract.
+            return Err(StdError::generic_err("input too long for addr_validate"));
+        }
+        let source = build_region(input_bytes);
         let source_ptr = &*source as *const Region as u32;
 
         let result = unsafe { addr_validate(source_ptr) };
@@ -169,11 +176,20 @@ impl Api for ExternalApi {
             )));
         }
 
-        Ok(Addr::unchecked(human))
+        Ok(Addr::unchecked(input))
     }
 
-    fn addr_canonicalize(&self, human: &str) -> StdResult<CanonicalAddr> {
-        let send = build_region(human.as_bytes());
+    fn addr_canonicalize(&self, input: &str) -> StdResult<CanonicalAddr> {
+        let input_bytes = input.as_bytes();
+        if input_bytes.len() > 90 {
+            // See MAX_LENGTH_HUMAN_ADDRESS in the VM.
+            // In this case, the VM will refuse to read the input from the contract.
+            // Stop here to allow handling the error in the contract.
+            return Err(StdError::generic_err(
+                "input too long for addr_canonicalize",
+            ));
+        }
+        let send = build_region(input_bytes);
         let send_ptr = &*send as *const Region as u32;
         let canon = alloc(CANONICAL_ADDRESS_BUFFER_LENGTH);